+++ source/Classes/00.Attributes/Documentation.ps1
@@ -1,0 +1,4 @@
+class Documentation : System.Attribute
+{
+    # Documentation() attribute used to specify type used for an API endpoint as output or input.
+}